refactor(claude): split daily and path loading
Move Claude daily summary loading and Claude path discovery into dedicated modules while keeping the primary entry loader in mod.rs.

This reduces the size of the adapter module without changing deduplication, date filtering, or project/session path extraction behavior.
