+++ rust/crates/ccusage/src/adapter/claude/mod.rs
@@ -1,6 +1,8 @@
+mod daily;
+mod paths;
+
 use std::{
-    collections::BTreeMap,
-    env, fs,
+    fs,
     hash::{Hash, Hasher},
     path::{Path, PathBuf},
     sync::Arc,
@@ -10,16 +12,21 @@ use std::{
 use jiff::tz::TimeZone as JiffTimeZone;
 use memchr::memmem;
 use rustc_hash::FxHasher;
-use serde::Deserialize;
 
 use crate::{
-    calculate_cost, calculate_cost_for_usage,
+    calculate_cost,
     cli::{CostMode, SharedArgs},
-    cli_error, debug_log,
-    fast::{byte_lines, suffix_string, FxHashMap, FxHashSet, SmallIndexVec},
-    format_date_tz, home, log_level, parse_ts_timestamp, parse_tz, progress, LoadedEntry,
-    LoadedFile, ModelBreakdown, PricingMap, Result, Speed, TimestampMs, TokenCounts, TokenUsageRaw,
-    UsageEntry, UsageSummary,
+    debug_log,
+    fast::{byte_lines, suffix_string, FxHashMap, SmallIndexVec},
+    format_date_tz, log_level, parse_ts_timestamp, parse_tz, progress, LoadedEntry, LoadedFile,
+    PricingMap, Result, Speed, TimestampMs, UsageEntry, UsageSummary,
+};
+
+#[cfg(test)]
+pub(crate) use paths::timestamp_from_line;
+pub(crate) use paths::{
+    claude_paths, collect_files_with_extension, collect_usage_files, extract_project,
+    extract_session_parts, usage_files,
 };
 
 pub(crate) fn load_entries(
@@ -37,84 +44,8 @@ pub(crate) fn load_daily_summaries(
     group_by_project: bool,
 ) -> Result<Vec<UsageSummary>> {
     progress::track_usage_load(progress::UsageLoadAgent::Claude, shared.json, || {
-        load_daily_summaries_inner(shared, project_filter, group_by_project)
+        daily::load_daily_summaries_inner(shared, project_filter, group_by_project)
     })
-}
-
-fn load_daily_summaries_inner(
-    shared: &SharedArgs,
-    project_filter: Option<&str>,
-    group_by_project: bool,
-) -> Result<Vec<UsageSummary>> {
-    let paths = claude_paths()?;
-    let files = usage_files(&paths, project_filter);
-    if files.is_empty() {
-        return Ok(Vec::new());
-    }
-
-    let pricing = if shared.mode == CostMode::Display {
-        None
-    } else {
-        Some(PricingMap::load(shared.offline, log_level() != Some(0)))
-    };
-    let tz = parse_tz(shared.timezone.as_deref());
-    let mode = shared.mode;
-    let loaded_files = if shared.single_thread {
-        files
-            .iter()
-            .map(|file| read_daily_usage_file(file, tz.as_ref(), mode, pricing.as_ref()))
-            .collect::<Vec<_>>()
-    } else {
-        read_daily_usage_files_parallel(&files, tz.as_ref(), mode, pricing.as_ref())
-    };
-
-    let mut deduped_indexes: FxHashMap<u64, SmallIndexVec> = FxHashMap::default();
-    let mut deduped = Vec::with_capacity(loaded_files.iter().map(|file| file.entries.len()).sum());
-    for loaded_file in loaded_files {
-        for entry in loaded_file.entries {
-            if let Some(filter) = project_filter {
-                if entry.project.as_ref() != filter {
-                    continue;
-                }
-            }
-            push_deduped_daily_entry(entry, &mut deduped_indexes, &mut deduped);
-        }
-    }
-
-    if group_by_project {
-        let mut groups = BTreeMap::<(String, Arc<str>), DailyAccumulator>::new();
-        for entry in &deduped {
-            groups
-                .entry((entry.date.clone(), Arc::clone(&entry.project)))
-                .or_default()
-                .add_entry(entry);
-        }
-        return Ok(groups
-            .into_iter()
-            .map(|((date, project), group)| {
-                let mut summary = group.into_summary();
-                summary.date = Some(date);
-                summary.project = Some(project.to_string());
-                summary
-            })
-            .collect());
-    }
-
-    let mut groups = BTreeMap::<String, DailyAccumulator>::new();
-    for entry in &deduped {
-        groups
-            .entry(entry.date.clone())
-            .or_default()
-            .add_entry(entry);
-    }
-    Ok(groups
-        .into_iter()
-        .map(|(key, group)| {
-            let mut summary = group.into_summary();
-            summary.date = Some(key);
-            summary
-        })
-        .collect())
 }
 
 fn load_entries_inner(
@@ -276,256 +207,12 @@ fn read_usage_files_parallel(
     })
 }
 
-#[derive(Debug)]
-struct DailyLoadedFile {
-    timestamp: Option<TimestampMs>,
-    entries: Vec<DailyLoadedEntry>,
-}
-
-#[derive(Debug)]
-struct DailyLoadedEntry {
-    date: String,
-    project: Arc<str>,
-    usage: TokenUsageRaw,
-    cost: f64,
-    model: Option<String>,
-    message_id: Option<String>,
-    request_id: Option<String>,
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct DailyUsageEntry {
-    timestamp: String,
-    message: DailyUsageMessage,
-    version: Option<String>,
-    session_id: Option<String>,
-    #[serde(rename = "costUSD")]
-    cost_usd: Option<f64>,
-    request_id: Option<String>,
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(untagged)]
-enum DailyUsageLine {
-    Direct(DailyUsageEntry),
-    AgentProgress(DailyAgentProgressEntry),
-}
-
-impl DailyUsageLine {
-    fn into_entry(self) -> DailyUsageEntry {
-        match self {
-            DailyUsageLine::Direct(entry) => entry,
-            DailyUsageLine::AgentProgress(entry) => DailyUsageEntry {
-                timestamp: entry.data.message.timestamp,
-                message: entry.data.message.message,
-                version: None,
-                session_id: None,
-                cost_usd: entry.data.message.cost_usd,
-                request_id: entry.data.message.request_id,
-            },
-        }
-    }
-}
-
-#[derive(Debug, Deserialize)]
-struct DailyAgentProgressEntry {
-    data: DailyAgentProgressData,
-}
-
-#[derive(Debug, Deserialize)]
-struct DailyAgentProgressData {
-    message: DailyAgentProgressMessage,
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct DailyAgentProgressMessage {
-    timestamp: String,
-    message: DailyUsageMessage,
-    #[serde(rename = "costUSD")]
-    cost_usd: Option<f64>,
-    request_id: Option<String>,
-}
-
-#[derive(Debug, Deserialize)]
-struct DailyUsageMessage {
-    usage: TokenUsageRaw,
-    model: Option<String>,
-    id: Option<String>,
-}
-
-fn read_daily_usage_files_parallel(
-    files: &[PathBuf],
-    tz: Option<&JiffTimeZone>,
-    mode: CostMode,
-    pricing: Option<&PricingMap>,
-) -> Vec<DailyLoadedFile> {
-    let worker_count = thread::available_parallelism()
-        .map(usize::from)
-        .unwrap_or(1)
-        .min(files.len());
-    if worker_count <= 1 {
-        return files
-            .iter()
-            .map(|file| read_daily_usage_file(file, tz, mode, pricing))
-            .collect();
-    }
-
-    let chunks = chunk_file_indexes_by_size(files, worker_count);
-    thread::scope(|scope| {
-        let mut handles = Vec::with_capacity(worker_count);
-        for chunk in chunks {
-            let tz = tz.cloned();
-            handles.push(scope.spawn(move || {
-                chunk
-                    .into_iter()
-                    .map(|index| {
-                        (
-                            index,
-                            read_daily_usage_file(&files[index], tz.as_ref(), mode, pricing),
-                        )
-                    })
-                    .collect::<Vec<_>>()
-            }));
-        }
-        let mut loaded_files = Vec::with_capacity(files.len());
-        loaded_files.resize_with(files.len(), || None);
-        for (index, file) in handles
-            .into_iter()
-            .flat_map(|handle| handle.join().expect("daily usage worker panicked"))
-        {
-            loaded_files[index] = Some(file);
-        }
-        loaded_files
-            .into_iter()
-            .map(|file| file.expect("daily usage worker returned every file"))
-            .collect()
-    })
-}
-
-fn read_daily_usage_file(
-    path: &Path,
-    tz: Option<&JiffTimeZone>,
-    mode: CostMode,
-    pricing: Option<&PricingMap>,
-) -> DailyLoadedFile {
-    let project: Arc<str> = Arc::from(extract_project(path));
-    let mut loaded_file = DailyLoadedFile {
-        timestamp: None,
-        entries: Vec::new(),
-    };
-    let Ok(content) = fs::read(path) else {
-        return loaded_file;
-    };
-
-    let usage_marker = memmem::Finder::new(br#""usage":{"#);
-    for line in byte_lines(&content) {
-        if usage_marker.find(line).is_none() {
-            continue;
-        }
-        if has_unsupported_null_field(line) {
-            continue;
-        }
-        let Ok(data) = serde_json::from_slice::<DailyUsageLine>(line) else {
-            continue;
-        };
-        let data = data.into_entry();
-        let Some(timestamp) = parse_ts_timestamp(&data.timestamp) else {
-            continue;
-        };
-        loaded_file.timestamp = Some(
-            loaded_file
-                .timestamp
-                .map_or(timestamp, |current| current.min(timestamp)),
-        );
-        if !is_valid_daily_usage_entry(&data) {
-            continue;
-        }
-        let usage = data.message.usage;
-        let cost = calculate_cost_for_usage(
-            data.message.model.as_deref(),
-            usage,
-            data.cost_usd,
-            mode,
-            pricing,
-        );
-        let model = data.message.model.as_ref().and_then(|model| {
-            if model == "<synthetic>" {
-                None
-            } else if matches!(usage.speed, Some(Speed::Fast)) {
-                Some(suffix_string(model, "-fast"))
-            } else {
-                Some(model.clone())
-            }
-        });
-        loaded_file.entries.push(DailyLoadedEntry {
-            date: format_date_tz(timestamp, tz),
-            project: Arc::clone(&project),
-            usage,
-            cost,
-            model,
-            message_id: data.message.id,
-            request_id: data.request_id,
-        });
-    }
-    loaded_file
-}
-
-fn is_valid_daily_usage_entry(data: &DailyUsageEntry) -> bool {
-    if data
-        .version
-        .as_deref()
-        .is_some_and(|version| !is_semver_prefix(version))
-    {
-        return false;
-    }
-    if data
-        .session_id
-        .as_deref()
-        .is_some_and(|session_id| session_id.is_empty())
-    {
-        return false;
-    }
-    if data
-        .request_id
-        .as_deref()
-        .is_some_and(|request_id| request_id.is_empty())
-    {
-        return false;
-    }
-    if data
-        .message
-        .id
-        .as_deref()
-        .is_some_and(|message_id| message_id.is_empty())
-    {
-        return false;
-    }
-    if data
-        .message
-        .model
-        .as_deref()
-        .is_some_and(|model| model.is_empty())
-    {
-        return false;
-    }
-    true
-}
-
 fn usage_token_total(data: &UsageEntry) -> u64 {
     let usage = data.message.usage;
     usage.input_tokens
         + usage.output_tokens
         + usage.cache_creation_input_tokens
         + usage.cache_read_input_tokens
-}
-
-fn daily_usage_token_total(entry: &DailyLoadedEntry) -> u64 {
-    entry.usage.input_tokens
-        + entry.usage.output_tokens
-        + entry.usage.cache_creation_input_tokens
-        + entry.usage.cache_read_input_tokens
 }
 
 fn should_replace_deduped_entry(candidate: &UsageEntry, existing: &UsageEntry) -> bool {
@@ -568,46 +255,6 @@ fn push_deduped_entry(
     }
 }
 
-fn push_deduped_daily_entry(
-    entry: DailyLoadedEntry,
-    deduped_indexes: &mut FxHashMap<u64, SmallIndexVec>,
-    deduped: &mut Vec<DailyLoadedEntry>,
-) {
-    let dedupe_lookup = entry.message_id.as_deref().map(|message_id| {
-        let request_id = entry.request_id.as_deref();
-        let hash = usage_dedupe_hash(message_id, request_id);
-        let existing_index = deduped_indexes.get(&hash).and_then(|indexes| {
-            indexes.iter().copied().find(|&index| {
-                deduped[index].message_id.as_deref() == Some(message_id)
-                    && deduped[index].request_id.as_deref() == request_id
-            })
-        });
-        (hash, existing_index)
-    });
-
-    if let Some((_, Some(index))) = dedupe_lookup {
-        let candidate_total = daily_usage_token_total(&entry);
-        let existing_total = daily_usage_token_total(&deduped[index]);
-        let should_replace = if candidate_total != existing_total {
-            candidate_total > existing_total
-        } else if entry.cost != deduped[index].cost {
-            entry.cost > deduped[index].cost
-        } else {
-            entry.usage.speed.is_some() && deduped[index].usage.speed.is_none()
-        };
-        if should_replace {
-            deduped[index] = entry;
-        }
-        return;
-    }
-
-    let index = deduped.len();
-    deduped.push(entry);
-    if let Some((hash, None)) = dedupe_lookup {
-        deduped_indexes.entry(hash).or_default().push(index);
-    }
-}
-
 fn usage_dedupe_hash(message_id: &str, request_id: Option<&str>) -> u64 {
     let mut hasher = FxHasher::default();
     message_id.hash(&mut hasher);
@@ -622,66 +269,6 @@ fn loaded_entry_matches_dedupe_key(
 ) -> bool {
     entry.data.message.id.as_deref() == Some(message_id)
         && entry.data.request_id.as_deref() == request_id
-}
-
-#[derive(Default)]
-struct DailyAccumulator {
-    counts: TokenCounts,
-    cost: f64,
-    models: Vec<String>,
-    breakdowns: Vec<ModelBreakdown>,
-    breakdown_indexes: FxHashMap<String, usize>,
-}
-
-impl DailyAccumulator {
-    fn add_entry(&mut self, entry: &DailyLoadedEntry) {
-        self.counts.add_usage(entry.usage);
-        self.cost += entry.cost;
-        if let Some(model) = &entry.model {
-            let index = if let Some(index) = self.breakdown_indexes.get(model.as_str()) {
-                *index
-            } else {
-                let index = self.breakdowns.len();
-                self.breakdown_indexes.insert(model.clone(), index);
-                self.models.push(model.clone());
-                self.breakdowns.push(ModelBreakdown {
-                    model_name: model.clone(),
-                    ..ModelBreakdown::default()
-                });
-                index
-            };
-            let breakdown = &mut self.breakdowns[index];
-            breakdown.input_tokens += entry.usage.input_tokens;
-            breakdown.output_tokens += entry.usage.output_tokens;
-            breakdown.cache_creation_tokens += entry.usage.cache_creation_input_tokens;
-            breakdown.cache_read_tokens += entry.usage.cache_read_input_tokens;
-            breakdown.cost += entry.cost;
-        }
-    }
-
-    fn into_summary(mut self) -> UsageSummary {
-        self.breakdowns.sort_by(|a, b| b.cost.total_cmp(&a.cost));
-        UsageSummary {
-            date: None,
-            month: None,
-            week: None,
-            session_id: None,
-            project_path: None,
-            last_activity: None,
-            input_tokens: self.counts.input_tokens,
-            output_tokens: self.counts.output_tokens,
-            cache_creation_tokens: self.counts.cache_creation_tokens,
-            cache_read_tokens: self.counts.cache_read_tokens,
-            extra_total_tokens: 0,
-            total_cost: self.cost,
-            credits: None,
-            message_count: None,
-            models_used: self.models,
-            model_breakdowns: self.breakdowns,
-            project: None,
-            versions: None,
-        }
-    }
 }
 
 fn read_usage_file(
@@ -759,7 +346,7 @@ fn update_loaded_file_timestamp(loaded_file: &mut LoadedFile, timestamp: Timesta
     );
 }
 
-fn is_valid_usage_entry(data: &UsageEntry) -> bool {
+pub(super) fn is_valid_usage_entry(data: &UsageEntry) -> bool {
     if data
         .version
         .as_deref()
@@ -800,7 +387,7 @@ fn is_valid_usage_entry(data: &UsageEntry) -> bool {
     true
 }
 
-fn has_unsupported_null_field(line: &[u8]) -> bool {
+pub(super) fn has_unsupported_null_field(line: &[u8]) -> bool {
     let mut offset = 0;
     while let Some(relative_index) = memmem::find(&line[offset..], b":null") {
         let null_index = offset + relative_index;
@@ -844,7 +431,7 @@ fn is_unsupported_nullable_field(field: &[u8]) -> bool {
     std::str::from_utf8(field).is_ok_and(|field| FIELDS.contains(field))
 }
 
-fn is_semver_prefix(value: &str) -> bool {
+pub(super) fn is_semver_prefix(value: &str) -> bool {
     let bytes = value.as_bytes();
     let mut index = 0;
     if !consume_ascii_digits(bytes, &mut index) || bytes.get(index) != Some(&b'.') {
@@ -864,189 +451,6 @@ fn consume_ascii_digits(bytes: &[u8], index: &mut usize) -> bool {
         *index += 1;
     }
     *index > start
-}
-
-pub(crate) fn claude_paths() -> Result<Vec<PathBuf>> {
-    let mut paths = Vec::new();
-    let mut seen = FxHashSet::default();
-    if let Ok(env_paths) = env::var("CLAUDE_CONFIG_DIR") {
-        for raw in env_paths
-            .split(',')
-            .map(str::trim)
-            .filter(|path| !path.is_empty())
-        {
-            let path = normalize_claude_config_path(raw);
-            if path.join("projects").is_dir() && seen.insert(path.clone()) {
-                paths.push(path);
-            }
-        }
-        if !paths.is_empty() {
-            return Ok(paths);
-        }
-        return Err(cli_error(format!(
-            "No valid Claude data directories found in CLAUDE_CONFIG_DIR. Expected each path to be a Claude config directory containing 'projects/', or the 'projects/' directory itself: {env_paths}"
-        )));
-    }
-
-    let home = home::home_dir().ok_or_else(|| cli_error("home directory is not set"))?;
-    let xdg = env::var("XDG_CONFIG_HOME")
-        .map(PathBuf::from)
-        .unwrap_or_else(|_| PathBuf::from(&home).join(".config"));
-    for path in [xdg.join("claude"), home.join(".claude")] {
-        if path.join("projects").is_dir() && seen.insert(path.clone()) {
-            paths.push(path);
-        }
-    }
-    if paths.is_empty() {
-        return Err(cli_error("No valid Claude data directories found"));
-    }
-    Ok(paths)
-}
-
-fn normalize_claude_config_path(raw: &str) -> PathBuf {
-    let path = expand_home_path(raw);
-    if path.file_name().is_some_and(|name| name == "projects") && path.is_dir() {
-        return path.parent().map(Path::to_path_buf).unwrap_or(path);
-    }
-    path
-}
-
-fn expand_home_path(raw: &str) -> PathBuf {
-    if raw == "~" {
-        if let Some(home) = home::home_dir() {
-            return home;
-        }
-    }
-    if let Some(rest) = raw.strip_prefix("~/") {
-        if let Some(home) = home::home_dir() {
-            return home.join(rest);
-        }
-    }
-    PathBuf::from(raw)
-}
-
-pub(crate) fn usage_files(paths: &[PathBuf], project_filter: Option<&str>) -> Vec<PathBuf> {
-    let mut files = Vec::new();
-    for path in paths {
-        let projects_dir = path.join("projects");
-        if let Some(project_filter) =
-            project_filter.filter(|filter| is_project_path_segment(filter))
-        {
-            collect_usage_files(&projects_dir.join(project_filter), &mut files);
-        } else {
-            collect_usage_files(&projects_dir, &mut files);
-        }
-    }
-    files.sort_by_cached_key(|path| path.to_string_lossy().into_owned());
-    files
-}
-
-fn is_project_path_segment(value: &str) -> bool {
-    !value.is_empty()
-        && value != "."
-        && value != ".."
-        && !value.contains('/')
-        && !value.contains('\\')
-}
-
-pub(crate) fn collect_usage_files(dir: &Path, files: &mut Vec<PathBuf>) {
-    collect_files_with_extension(dir, "jsonl", files);
-}
-
-pub(crate) fn collect_files_with_extension(dir: &Path, extension: &str, files: &mut Vec<PathBuf>) {
-    let Ok(entries) = fs::read_dir(dir) else {
-        return;
-    };
-
-    for entry in entries.filter_map(std::result::Result::ok) {
-        let Ok(file_type) = entry.file_type() else {
-            continue;
-        };
-        let path = entry.path();
-        if file_type.is_file() && path.extension().is_some_and(|ext| ext == extension) {
-            files.push(path);
-        } else if file_type.is_dir() {
-            collect_files_with_extension(&path, extension, files);
-        }
-    }
-}
-
-#[cfg(test)]
-pub(crate) fn timestamp_from_line(line: &str) -> Option<TimestampMs> {
-    timestamp_from_line_bytes(line.as_bytes())
-}
-
-#[cfg(test)]
-fn timestamp_from_line_bytes(line: &[u8]) -> Option<TimestampMs> {
-    let marker = br#""timestamp":""#;
-    let start = memmem::find(line, marker)? + marker.len();
-    let end = memchr::memchr(b'"', &line[start..])? + start;
-    let timestamp = std::str::from_utf8(&line[start..end]).ok()?;
-    parse_ts_timestamp(timestamp)
-}
-
-pub(crate) fn extract_project(path: &Path) -> String {
-    let mut saw_projects = false;
-    for part in path
-        .components()
-        .filter_map(|component| component.as_os_str().to_str())
-    {
-        if saw_projects {
-            return if part.trim().is_empty() {
-                "unknown"
-            } else {
-                part
-            }
-            .to_string();
-        }
-        if part == "projects" {
-            saw_projects = true;
-        }
-    }
-    "unknown".to_string()
-}
-
-pub(crate) fn extract_session_parts(path: &Path) -> (String, String) {
-    let parts = path
-        .components()
-        .filter_map(|component| component.as_os_str().to_str())
-        .collect::<Vec<_>>();
-    let projects_index = parts.iter().position(|part| *part == "projects");
-    let relative = projects_index
-        .map(|index| &parts[index + 1..])
-        .unwrap_or(&parts);
-    let file_session_id = relative
-        .last()
-        .and_then(|file_name| file_name.strip_suffix(".jsonl"))
-        .filter(|session_id| !session_id.is_empty());
-    if relative.len() == 2 {
-        if let Some(session_id) = file_session_id {
-            return (session_id.to_string(), relative[0].to_string());
-        }
-    }
-    if relative.len() >= 4 && relative.get(relative.len() - 2) == Some(&"subagents") {
-        let session_id = relative[relative.len() - 3].to_string();
-        let project_path = relative[..relative.len() - 3].join(std::path::MAIN_SEPARATOR_STR);
-        return (
-            session_id,
-            if project_path.is_empty() {
-                "Unknown Project".to_string()
-            } else {
-                project_path
-            },
-        );
-    }
-    let session_id = relative
-        .get(relative.len().saturating_sub(2))
-        .copied()
-        .unwrap_or("unknown")
-        .to_string();
-    let project_path = if relative.len() > 2 {
-        relative[..relative.len() - 2].join(std::path::MAIN_SEPARATOR_STR)
-    } else {
-        "Unknown Project".to_string()
-    };
-    (session_id, project_path)
 }
 
 #[cfg(test)]
@@ -1089,7 +493,8 @@ mod tests {
     use std::{fs, path::Path};
 
     use super::{
-        extract_session_parts, has_unsupported_null_field, is_project_path_segment, usage_files,
+        extract_session_parts, has_unsupported_null_field, paths::is_project_path_segment,
+        usage_files,
     };
 
     fn temp_claude_dir(name: &str) -> std::path::PathBuf {
